refactor(watcher): extract _websocket_logic (462 lines) into watcher_ws_logic.py
watcher.py is a god class. This commit extracts the largest
async method remaining in the orchestrator: a single WebSocket
connection lifecycle (dial, auth, heartbeat, receive, fallback).

- New: src/gengowatcher/watcher_ws_logic.py (499 lines)
  Exports:
    * async websocket_logic(watcher)
      Drives one WebSocket session: reads wss_url, sets
      websocket_status='Connecting', calls
      asyncio.to_thread(watcher._sync_session_before_websocket_connect),
      builds the auth payload via build_websocket_auth_payload,
      negotiates an additional-headers set from
      build_browser_aligned_websocket_headers, dials
      websockets.connect with open_timeout=20, ping_interval=20,
      ping_timeout=10, sends the auth JSON, and runs four concurrent
      tasks: heartbeat, refresh_session_token_periodically,
      monitor_test_request, and the main receive loop. On handshake
      failures it retries without the custom headers. On clean
      closes (1000/1001) it sets websocket_status='Offline'; on
      abnormal closes / ConnectionClosed / InvalidStatusCode /
      InvalidHandshake it logs and exits the inner loop so the
      outer watcher_ws_monitor.run_websocket_monitor can apply
      backoff. Receives 'available_collection' events and forwards
      them to watcher._process_new_job with the parsed title/reward
      /url/etc.

  The transformation was mechanical: dedent by 4 spaces, rewrite
  the signature async def _websocket_logic(self) to
  async def websocket_logic(watcher), and rename every
  self. to watcher.. The nested async def heartbeat(self)
  inner coroutine was renamed in the same way. All async-control
  flow is preserved verbatim.

  Adds imports for websockets.exceptions (ConnectionClosed /
  InvalidHandshake / InvalidStatusCode) and browser_session
  helpers (build_browser_aligned_websocket_headers,
  build_websocket_auth_payload, format_cookies_as_header).

- watcher.py:
  * GengoWatcher._websocket_logic is now a thin async delegator:
    return await _websocket_logic_impl(self).
  * Dropped the unused 'import json', the
    'from websockets.exceptions import ...' block, and the now-
    unused .browser_session names (build_browser_aligned_websocket_headers,
    build_websocket_auth_payload, format_cookies_as_header) from
    the module-level imports.

The single call site from watcher_ws_monitor.run_websocket_monitor
(asyncio.run(watcher._websocket_logic())) continues to resolve
through the class, as do the test sites in tests/test_websocket.py
(108, 156) and tests/test_watcher_comprehensive.py:1253.

Size delta: watcher.py 2258 -> 1796 lines (-462 LoC). The 462-line
async WebSocket logic now lives in watcher_ws_logic.py.

Verification:
  ruff check src/ tests/   -> All checks passed
  pytest -q                -> 760 passed, 4 skipped
